requirements.txt
pymodbus==3.10.0
loguru==0.7.3
pytest==8.4.1
python-dotenv==1.1.1
colorama==0.4.6

